+++ database/migrations/2020_03_02_215003_create_refunds_table.php
@@ -14,10 +14,10 @@ class CreateRefundsTable extends Migration
     public function up()
     {
         Schema::create('refunds', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->increments('id');
 
-            $table->unsignedBigInteger('payment_id')->nullable();
-            $table->unsignedBigInteger('order_id')->nullable();
+            $table->unsignedInteger('payment_id')->nullable();
+            $table->unsignedInteger('order_id')->nullable();
             $table->json('price');
             $table->unsignedSmallInteger('user_id')->nullable();
             $table->unsignedTinyInteger('refund_type_id')->nullable();
